AI #909: SkuPriceDetails req. fix -  prefer FOCUS-defined over Provider-defined properties (#910)
The following requirement does not clearly define the precedence between
FOCUS-defined and Provider-defined properties. This could lead to
ambiguity or duplication when both exist.

> * SkuPriceDetails MUST include the FOCUS-defined SKU Price property
when an equivalent property is included as a Provider-defined property.

The goal of this PR is to clarify that FOCUS-defined properties should
be preferred and used instead of equivalent Provider-defined ones.

---------

Co-authored-by: Michael Flanakin <flanakin@users.noreply.github.com>