FR #1090: Associate attributes with datasets (#1431)
We need to finesse the spec to associate specific attributes with
specific datasets.

Current verbiage is:

> Attributes are requirements that apply across all FOCUS datasets
instead of an individual column level.

1) That's not true for all attributes: we apply some at the column
level.
2) This means that every attribute is associated with every dataset,
which will be problematic/inaccurate as our number of datasets grows
beyond one. For example, Discount Handling has nothing to do with
Contracts.

This PR therefore associates the `Cost and Usage` dataset to all the
current attributes.

FYI that PR #1350 performs the same association for `Contract
Commitment` (or rather, it will, once we get agreement on this PR).

---------

Co-authored-by: Irena Jurica <irena.jurica@neos.hr>