FR #1070: Clarify use of "period" vs "duration" vs "term" (#1493)
When authoring the spec, we have been inconsistent in our use of the
following words:

- `period`
- `duration`
- `term`

We have had a glossary entry for `term` that means "the length of a
commitment discount". However, with the introduction of Contract
Commitments via PR #1350, it becomes important for the word "term" to
not just mean length of time, but also any particular parameter
specified on the contract. **Thus, any legacy use of `term` in the FOCUS
spec when describing time becomes potentially confusing and/or
inaccurate.**

Therefore, this consistency PR performs the following changes:

- Add glossary entry for `period`, which is simply defined as "a length
of time".
- **Change all previous use of `term` and `duration` to `period` (or
`commitment period`, when applicable) when used in the context of
*time*.**
- Strike `term` from the glossary.

As a consistency PR, this should not introduce any material change to
the normative requirements of the specification.