FRs #1055/1073: Refactor Provider Identification, Host Provider, Invoice Issuer (#1139)
We've received feedback from multiple sources indicating that our column
names and definitions for Provider and Publisher are confusing.
Additionally, we want to add support for vendors that host their systems
on other providers, as reporting this information is important for cost
analysis.

As such are introducing the following changes:

- (Deprecate) Publisher Name: Publisher Name is not specific enough, so
we're deprecating this column, as the new and previously introduced
columns provide the needed specificity.
- (Deprecate): Provider Name: Provider Name was originally planned to be
renamed to Service Provider; however, based on feedback, we've elected
to deprecate Provider and introduce the new Service Provider column
- (New Column) Service Provider with the same definition as Provider
- (New Column) Host Provider has been introduced to support our new
feature, which allows Data Generators to provide hosting information to
their customers.

Invoice Issuer Name and Data Generator Name will continue unchanged.

We have also update examples to indicate different entity scenarios. 

We believe these changes will resolve the confusion and help
practitioners better understand the entities involved in the FinOps
discipline.

For an explanation of how deprecations are handled see
https://github.com/FinOps-Open-Cost-and-Usage-Spec/FOCUS_Spec/blob/working_draft/guidelines/spec-change-guidelines.md

---------

Co-authored-by: Shawn Alpay <77511110+shawnalpay@users.noreply.github.com>
Co-authored-by: Irena Jurica <irena.jurica@neos.hr>
Co-authored-by: Michael Flanakin <flanakin@users.noreply.github.com>
Co-authored-by: Udam Dewaraja <udam@stitcher.ai>
Co-authored-by: Joaquin <jprado@linuxfoundation.org>